pyproject.toml
[project]
name = "python-agentframework-demos"
version = "0.1.0"
description = "Collection of Python examples for Microsoft Agent Framework"
requires-python = ">=3.10"
dependencies = [
    "azure-identity",
    "openai>=1.109.1",
    "python-dotenv",
    "pydantic",
    "rich",
    "dotenv-azd",
    "aiohttp",
    "faker",
    "fastmcp",
    "opentelemetry-exporter-otlp-proto-grpc",
    "azure-monitor-opentelemetry",
    "psycopg[binary]",
    "pgvector",
    "markitdown[pdf]",
    "azure-ai-evaluation[redteam]>=1.15.0",
    "agent-framework-core==1.0.0",
    "agent-framework-openai==1.0.0",
    "agent-framework-devui==1.0.0b260402",
    "agent-framework-redis==1.0.0b260402",
    "agent-framework-mem0==1.0.0b260402",
    "agent-framework-azure-ai-search==1.0.0b260402",
    "agent-framework-orchestrations==1.0.0b260402",
    "agent-framework-foundry>=1.0.0",
]

[tool.uv]
prerelease = "allow"

[tool.ruff]
line-length = 120
target-version = "py310"
lint.select = ["E", "F", "I", "UP"]
lint.ignore = ["D203"]
